+++ apps/ledger-live-mobile/src/screens/FirmwareUpdate/index.tsx
@@ -48,6 +48,7 @@ import { targetDataDimensions } from "../CustomImage/shared";
 import { ProcessorPreviewResult } from "../../components/CustomImage/ImageProcessor";
 import { ImageSourceContext } from "../../components/CustomImage/StaxFramedImage";
 import Button from "../../components/wrappedUi/Button";
+import Link from "../../components/wrappedUi/Link";
 import { RestoreStepDenied } from "./RestoreStepDenied";
 
 type FirmwareUpdateProps = {
@@ -469,18 +470,32 @@ export const FirmwareUpdate = ({
           <Button
             event="button_clicked"
             eventProperties={{
-              button: "Quit flow",
+              button: "Retry flow",
               screen: "Firmware update",
               drawer: `Error: ${error.name}`,
             }}
             type="main"
             outline={false}
-            onPress={quitUpdate}
+            onPress={retryCurrentStep}
             mt={6}
             alignSelf="stretch"
           >
-            {t("FirmwareUpdate.quitUpdate")}
+            {t("common.retry")}
           </Button>
+          <Flex mt={7} alignSelf="stretch">
+            <Link
+              event="button_clicked"
+              eventProperties={{
+                button: "Quit flow",
+                screen: "Firmware update",
+                drawer: `Error: ${error.name}`,
+              }}
+              type="main"
+              onPress={quitUpdate}
+            >
+              {t("FirmwareUpdate.quitUpdate")}
+            </Link>
+          </Flex>
         </DeviceActionError>
       );
     }
@@ -541,15 +556,36 @@ export const FirmwareUpdate = ({
     }
 
     if (staxLoadImageState.imageLoadRequested) {
-      return renderImageLoadRequested({ t, device, fullScreen: false });
+      return renderImageLoadRequested({
+        t,
+        device,
+        fullScreen: false,
+        wording: t("FirmwareUpdate.steps.restoreSettings.imageLoadRequested", {
+          deviceName: productName,
+        }),
+      });
     }
 
     if (staxLoadImageState.imageCommitRequested) {
-      return renderImageCommitRequested({ t, device, fullScreen: false });
+      return renderImageCommitRequested({
+        t,
+        device,
+        fullScreen: false,
+        wording: t("FirmwareUpdate.steps.restoreSettings.imageCommitRequested", {
+          deviceName: productName,
+        }),
+      });
     }
 
     if (restoreAppsState.allowManagerRequestedWording) {
-      return <AllowManager device={device} wording={t("DeviceAction.allowSecureConnection")} />;
+      return (
+        <AllowManager
+          device={device}
+          wording={t("FirmwareUpdate.steps.restoreSettings.allowAppsRestoration", {
+            deviceName: productName,
+          })}
+        />
+      );
     }
 
     if (installLanguageState.languageInstallationRequested) {
@@ -558,6 +594,9 @@ export const FirmwareUpdate = ({
         device,
         theme,
         fullScreen: false,
+        wording: t("FirmwareUpdate.steps.restoreSettings.allowLanguageInstallation", {
+          deviceName: productName,
+        }),
       });
     }
 
@@ -578,21 +617,22 @@ export const FirmwareUpdate = ({
     updateActionState.error,
     updateActionState.step,
     updateActionState.progress,
+    deviceLockedOrUnresponsive,
+    hasReconnectErrors,
     staxLoadImageState.imageLoadRequested,
     staxLoadImageState.imageCommitRequested,
-    installLanguageState.languageInstallationRequested,
     restoreAppsState.allowManagerRequestedWording,
+    installLanguageState.languageInstallationRequested,
+    restoreStepDeniedError,
     device,
     t,
-    theme,
+    retryCurrentStep,
     quitUpdate,
     deviceInfo.seVersion,
     firmwareUpdateContext.final.name,
     firmwareUpdateContext.shouldFlashMCU,
-    retryCurrentStep,
-    hasReconnectErrors,
-    deviceLockedOrUnresponsive,
-    restoreStepDeniedError,
+    theme,
+    productName,
     skipCurrentRestoreStep,
   ]);