+++ apps/ledger-live-mobile/src/screens/Swap/Form/index.tsx
@@ -143,6 +143,8 @@ export function SwapForm({
     onBeforeTransaction,
     excludeFixedRates: true,
     providers,
+    timeout: 10000,
+    timeoutErrorMessage: t("errors.SwapTimeoutError.title"),
   });
 
   const exchangeRatesState = swapTransaction.swap?.rates;