+++ libs/ledgerjs/packages/types-live/src/feature.ts
@@ -12,6 +12,7 @@ export type FeatureId =
   | "ptxSmartRouting"
   | "ptxEarn"
   | "currencyAvalancheCChain"
+  | "currencyStacks"
   | "currencyOptimism"
   | "currencyOptimismGoerli"
   | "currencyArbitrum"
@@ -28,6 +29,10 @@ export type FeatureId =
   | "currencyVelasEvm"
   | "currencySyscoin"
   | "currencyAxelar"
+  | "currencySecretNetwork"
+  | "currencyDesmos"
+  | "currencyUmee"
+  | "currencyStargaze"
   | "currencyOnomy"
   | "currencyPersistence"
   | "currencyQuicksilver"
@@ -67,7 +72,8 @@ export type FeatureId =
   | "protectServicesDiscoverDesktop"
   | "protectServicesDesktop"
   | "transactionsAlerts"
-  | "listAppsV2";
+  | "listAppsV2"
+  | "editEthTx";
 
 /**  We use objects instead of direct booleans for potential future improvements
 like feature versioning etc */