+++ libs/ledgerjs/packages/hw-transport-http/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ledgerhq/hw-transport-http
 
+## 6.28.0
+
+### Minor Changes
+
+- [#3523](https://github.com/LedgerHQ/ledger-live/pull/3523) [`500e36a789`](https://github.com/LedgerHQ/ledger-live/commit/500e36a789b34125164433fbd03c00305669ee4d) Thanks [@drakoFukayu](https://github.com/drakoFukayu)! - feat: LLD allow user to use vault as a signer
+
+## 6.28.0-next.0
+
+### Minor Changes
+
+- [#3523](https://github.com/LedgerHQ/ledger-live/pull/3523) [`500e36a789`](https://github.com/LedgerHQ/ledger-live/commit/500e36a789b34125164433fbd03c00305669ee4d) Thanks [@drakoFukayu](https://github.com/drakoFukayu)! - feat: LLD allow user to use vault as a signer
+
 ## 6.27.16
 
 ### Patch Changes