+++ libs/live-network/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/live-network
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`5af41b6fa1`](https://github.com/LedgerHQ/ledger-live/commit/5af41b6fa1e43037ccdb2df279c82e12ef3d2b1a)]:
+  - @ledgerhq/live-env@0.3.1
+
+## 1.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`5af41b6fa1`](https://github.com/LedgerHQ/ledger-live/commit/5af41b6fa1e43037ccdb2df279c82e12ef3d2b1a)]:
+  - @ledgerhq/live-env@0.3.1-next.0
+
 ## 1.1.1
 
 ### Patch Changes